backend/quiz-service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.1.1
python-dotenv==0.19.1
celery==5.2.3
redis==4.1.1
requests==2.26.0
werkzeug==2.0.1
google-generativeai==0.3.1
PyJWT==2.4.0
